refactor(db): extract filesystem utilities into separate module
Move the cross-platform home directory path retrieval function from
database.cpp and database_core.cpp into a new filesystem_utils module
to improve code organization and reusability. Update CMakeLists.txt
to include the new utility files in the build.
