+++ CMakeLists.txt
@@ -79,6 +79,9 @@ add_library(llm_core STATIC
     database/message_repository.h
     database/model_repository.cpp
     database/model_repository.h
+    # Utility modules
+    filesystem_utils.cpp
+    filesystem_utils.h
     tools.cpp
     tools.h
     model_manager.cpp
