+++ database.cpp
@@ -6,28 +6,6 @@
 #include <stdexcept>
 #include <optional>
 
-namespace { // Anonymous namespace for helper
-std::filesystem::path get_home_directory_path() {
-    #ifdef _WIN32
-        const char* userprofile = std::getenv("USERPROFILE");
-        if (userprofile) {
-            return std::filesystem::path(userprofile);
-        }
-        const char* homedrive = std::getenv("HOMEDRIVE");
-        const char* homepath = std::getenv("HOMEPATH");
-        if (homedrive && homepath) {
-            return std::filesystem::path(homedrive) / homepath;
-        }
-    #else // POSIX-like systems
-        const char* home_env = std::getenv("HOME");
-        if (home_env) {
-            return std::filesystem::path(home_env);
-        }
-    #endif
-    return ""; // Return empty path if home directory cannot be determined
-}
-} // end anonymous namespace
-
 namespace { // Or make these part of a utility struct/namespace if preferred
 struct SQLiteStmtDeleter {
     void operator()(sqlite3_stmt* stmt) const {
